requirements.txt
pyserial
opcua
time